+++ packages/@tinacms/icons/src/Exit.tsx
@@ -16,13 +16,17 @@ limitations under the License.
 
 */
 
-import { Schema } from 'prosemirror-model'
-import { MarkdownTranslator, Format, DOMTranslator } from '../../translator'
+import * as React from 'react'
 
-export const buildTranslator = (
-  schema: Schema,
-  format: Format = 'markdown'
-) => {
-  if (format === 'html') return DOMTranslator.fromSchema(schema)
-  return MarkdownTranslator.fromSchema(schema)
-}
+export const ExitIcon = () => (
+  <svg
+    width="32"
+    height="32"
+    viewBox="0 0 32 32"
+    fill="inherit"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M15.5 23.0129L8.88889 23.0129L8.88889 9.10324L15.5 9.10324L15.5 7.11615L8.88889 7.11615C7.85 7.11615 7 8.01034 7 9.10324L7 23.0129C7 24.1058 7.85 25 8.88889 25L15.5 25L15.5 23.0129Z" />
+    <path d="M18.6961 12.4912L21.1328 15.0645L12 15.0645L12 17.0516L21.1328 17.0516L18.6961 19.6249L20.0278 21.0258L24.75 16.0581L20.0278 11.0903L18.6961 12.4912Z" />
+  </svg>
+)