+++ packages/react-tinacms-editor/src/schema/marks/strong.ts
@@ -38,4 +38,5 @@ export const strong = {
   toDOM() {
     return ['strong']
   },
+  excludes: 'code',
 }