+++ packages/react-tinacms-inline/src/blocks/add-block-menu.tsx
@@ -22,16 +22,18 @@ import { BlockTemplate } from 'tinacms'
 import { IconButton } from '@tinacms/styles'
 import { AddIcon } from '@tinacms/icons'
 
+import { getOffset, getOffsetX, getOffsetY } from '../styles'
+
 interface AddBlockMenuProps {
   addBlock(data: any): void
-  templates: BlockTemplate[]
+  blocks: { [key: string]: { template: BlockTemplate } }
   position?: 'top' | 'bottom' | 'left' | 'right'
   index?: number
-  offset?: number
+  offset?: number | { x: number; y: number }
 }
 
 export function AddBlockMenu({
-  templates,
+  blocks,
   addBlock,
   position,
   index,
@@ -59,12 +61,15 @@ export function AddBlockMenu({
       }
     }
 
-    templates.length == 1
-      ? addBlock({
-          _template: templates[0].type,
-          ...templates[0].defaultItem,
-        })
-      : setIsOpen(isOpen => !isOpen)
+    if (Object.keys(blocks).length == 1) {
+      const blockId = Object.keys(blocks)[0]
+      addBlock({
+        _template: blockId,
+        ...blocks[blockId].template.defaultItem,
+      })
+    } else {
+      setIsOpen(isOpen => !isOpen)
+    }
   }
 
   React.useEffect(() => {
@@ -72,8 +77,6 @@ export function AddBlockMenu({
     document.addEventListener('mouseup', inactivateBlockMenu, false)
     return () => document.removeEventListener('mouseup', inactivateBlockMenu)
   }, [])
-
-  templates = templates || []
 
   return (
     <AddBlockWrapper
@@ -92,21 +95,24 @@ export function AddBlockMenu({
         <AddIcon />
       </AddBlockButton>
       <BlocksMenu openTop={openTop} isOpen={isOpen}>
-        {templates.map((template: BlockTemplate) => (
-          <BlockOption
-            key={template.label}
-            onClick={event => {
-              event.stopPropagation()
-              event.preventDefault()
-              addBlock({
-                _template: template.type,
-                ...template.defaultItem,
-              })
-            }}
-          >
-            {template.label}
-          </BlockOption>
-        ))}
+        {Object.keys(blocks).map((key: string) => {
+          const template = blocks[key].template
+          return (
+            <BlockOption
+              key={template.label}
+              onClick={event => {
+                event.stopPropagation()
+                event.preventDefault()
+                addBlock({
+                  _template: key,
+                  ...template.defaultItem,
+                })
+              }}
+            >
+              {template.label}
+            </BlockOption>
+          )
+        })}
       </BlocksMenu>
     </AddBlockWrapper>
   )
@@ -144,45 +150,48 @@ const AddBlockButton = styled(IconButton)<AddMenuProps>`
 
 interface AddBlockWrapperProps {
   index?: number
-  offset?: number
+  offset?: number | { x: number; y: number }
   position?: 'top' | 'bottom' | 'left' | 'right'
   isOpen: boolean
 }
 
-const AddBlockWrapper = styled.div<AddBlockWrapperProps>(
-  p => css`
+const AddBlockWrapper = styled.div<AddBlockWrapperProps>(p => {
+  const offset = getOffset(p.offset)
+
+  return css`
   position: absolute;
   z-index: calc(var(--tina-z-index-2) - ${p.index !== undefined ? p.index : 0});
 
-  ${p.position == 'top' &&
+
+  ${p.position === 'top' &&
     css`
-      top: calc(-1 * ${p.offset !== undefined ? p.offset : `16`}px);
+      top: calc(-1 * ${getOffsetY(offset)}px);
       left: 50%;
       transform: translate3d(-50%, -50%, 0);
     `}
 
-  ${p.position == 'left' &&
+  ${p.position === 'left' &&
     css`
       top: 50%;
-      left: calc(-1 * ${p.offset !== undefined ? p.offset : `16`}px);
+      left: calc(-1 * ${getOffsetX(offset)}px);
       transform: translate3d(-50%, -50%, 0);
     `}
 
-  ${p.position == 'bottom' &&
+  ${p.position === 'bottom' &&
     css`
-      bottom: calc(-1 * ${p.offset !== undefined ? p.offset : `16`}px);
+      bottom: calc(-1 * ${getOffsetY(offset)}px);
       left: 50%;
       transform: translate3d(-50%, 50%, 0);
     `}
 
-  ${p.position == 'right' &&
+  ${p.position === 'right' &&
     css`
       top: 50%;
-      right: calc(-1 * ${p.offset !== undefined ? p.offset : `16`}px);
+      right: calc(-1 * ${getOffsetX(offset)}px);
       transform: translate3d(50%, -50%, 0);
     `}
 
-  ${p.position == undefined &&
+  ${p.position === undefined &&
     css`
       position: relative;
     `}
@@ -192,7 +201,7 @@ const AddBlockWrapper = styled.div<AddBlockWrapperProps>(
       z-index: calc(1 + var(--tina-z-index-2) - ${p.index ? p.index : 0});
     `}
 `
-)
+})
 
 const BlocksMenu = styled.div<AddMenuProps>`
   min-width: 192px;