+++ packages/demo-next/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.5.0-alpha.0...nextjs-blog-starter@0.5.0) (2020-06-15)
+
+**Note:** Version bump only for package nextjs-blog-starter
+
+
+
+
+
+# [0.5.0-alpha.0](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.4.3-alpha.0...nextjs-blog-starter@0.5.0-alpha.0) (2020-06-12)
+
+
+### Features
+
+* allow a custom empty state component to be added ([abe9479](https://github.com/kendallstrautman/nextjs-blog-starter/commit/abe9479))
+
+
+
+
+
 ## [0.4.3-alpha.0](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.4.2...nextjs-blog-starter@0.4.3-alpha.0) (2020-06-08)
 
 **Note:** Version bump only for package nextjs-blog-starter